+++ src/ts/src/public/account/AccountApiUpdateRequest.ts
@@ -1,0 +1,9 @@
+import { AccountItemDto } from "./dto";
+
+export interface AccountApiUpdateRequest {
+  displayName?: string;
+}
+
+export interface AccountApiUpdateResponse {
+  item: AccountItemDto;
+}