+++ src/ts/src/public/account/index.ts
@@ -6,4 +6,5 @@ export * from "./AccountApiCreateRequest";
 export * from "./AccountApiLoginRequest";
 export * from "./AccountApiMigrateAuth0Request";
 export * from "./AccountApiPasswordChangeRequest";
+export * from "./AccountApiUpdateRequest";
 export * from "./AccountIsOnboardRequest";